sprint 1 fixes that Tianqi suggested and moved functions from routing into auth services file
