+++ app/src/main/kotlin/Auth_Services.kt
@@ -12,6 +12,15 @@ import io.ktor.server.pebble.PebbleContent
 import io.ktor.http.*
 import io.ktor.server.http.content.resources
 import io.ktor.server.http.content.static
+import java.io.File
+import org.mindrot.jbcrypt.BCrypt
+import org.jetbrains.exposed.v1.core.*
+import org.jetbrains.exposed.v1.jdbc.selectAll
+import org.jetbrains.exposed.v1.jdbc.transactions.transaction
+
+
+const val MAX_EMAIL_LENGTH = 128
+const val MIN_PASSWORD_LENGTH = 8
 
 suspend fun ApplicationCall.SignUpPage() {
     respondTemplate("pages/auth/signup.peb", model = emptyMap())
@@ -71,4 +80,24 @@ private suspend fun ApplicationCall.getCredentials(): Pair<String, String> {
     val email = parameters.getOrFail("email")
     val password = parameters.getOrFail("password")
     return email to password
+}
+
+fun isEmailValid(email: String): Boolean{
+    val preexsistingUser = UserDatabase.isEmailDuplicate(email)
+
+    return when {
+        preexsistingUser -> false
+        email.length < MAX_EMAIL_LENGTH -> true
+        email.length > 1 -> true
+        email.all { it.isLetterOrDigit() || it in setOf('@', '.', '_') } -> true
+        else -> false
+    }
+}
+
+fun hashPasswordIfValid(password: String): String? {
+    return if (password.length >= MIN_PASSWORD_LENGTH && password.all { !it.isWhitespace() }) {
+        BCrypt.hashpw(password, BCrypt.gensalt())
+    } else {
+        null
+    }
 }
